Move Directory extensions into the correct file
